+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ﻿# Autofocus Analysis
+## 2.0.0.2
+- Upgrade to N.I.N.A. Beta packages
+- Reduced the default temperature preset to -50 °C
+
+## 2.0.0.1
+- Upgrade to NET7 for N.I.N.A. 3.0
+
+## 1.0.0.1
+- Fix issue when Fitting was not using Hyperbolic with Hocus Focus
 
 ## 1.0.0.0
 - Initial release